AAAAAA

+++ 453505/ОАиП/LR_1/Task_3.cpp
@@ -2,7 +2,7 @@
 #include <cmath> // Для функции sin
 
 int main() {
-    double start, end, step;
+    int start, end, step;
 
     std::cout << "Введите начальное значение x: ";
     std::cin >> start; // Задаем исходное значение переменной
